% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMat.R
\name{covMat}
\alias{covMat}
\title{Stage 1 analysis collation of variance parameters}
\usage{
covMat(vc, which.comb, ltrait)
}
\arguments{
\item{vc}{a list of varcomp summary objects from each of the stage 1
analyses using \code{asreml}}

\item{which.comb}{A matrix of combinations of levels of the factor
used to subset the stage 1 analyses}

\item{ltrait}{the levels of the trait indexing the subsetting in
the stage 1 analyses}
}
\description{
Collate all variance parameters in a stage 1 analysis
}
\details{
Stage 1 analysis variance parameters
}
\author{
Ari Verbyla (averbyla at avdataanalytics.com.au)
}
