% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faSummary.R
\name{faSummary}
\alias{faSummary}
\title{faSummary}
\usage{
faSummary(fm, data, Trait, id = "id")
}
\arguments{
\item{fm}{a list of fitted models, from a diagonal form to factor analytic models
of increasing order}

\item{data}{the data frame used in fitting the models}

\item{Trait}{a character string giving the names of the trait factor (for example site)}

\item{id}{a character string giving the name of the genotype/variety factor.  Default is
"id"}
}
\value{
a data frame with various statistics and diagnostics.
}
\description{
faSummary
}
\details{
For a list of models from a diagonal form (always assumed first in the list),
to a sequential set of factor analysis (FA) models, tabulate various statistics
to provide guidance on model choice of an FA model.
}
