% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frobenius.R
\name{Frobenius}
\alias{Frobenius}
\title{Frobenius Norm}
\usage{
Frobenius(D)
}
\arguments{
\item{D}{a matrix}
}
\value{
the Frobenius norm, a scalar
}
\description{
Frobenius norm
}
\details{
This function calculates a distance measure using the Frobenius norm
}
