% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genherit1.default.R
\name{genherit1.default}
\alias{genherit1.default}
\title{genherit1 for method 'default'}
\usage{
\method{genherit1}{default}(asr, ...)
}
\arguments{
\item{asr}{a model object}

\item{\ldots}{Additional parameter}
}
\description{
Internal lmmtools function
}
\details{
Default method for calculating the generalised heritability for a single site or trait.
The only available method is 'asreml'.
}
