% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icREML.R
\name{icREML}
\alias{icREML}
\title{Find the AIC and BIC for a \code{list} of models fitted using
\code{asreml}}
\usage{
icREML(fm, scale = 1, logdet = FALSE)
}
\arguments{
\item{fm}{A \code{list} of asreml fitted model objects}

\item{scale}{A scalar to scale the variance matrix of the estimated
fixed effects (to ensure numerical stability of a log-determinant).
Default value is 1.}

\item{logdet}{The log determinant that modifies the residual
log-likelihood to form the full log-likelihood is to be included in the
table.}
}
\value{
A data frame.  The data frame has the following components
\itemize{
\item \code{model} : the names of the models
\item \code{res.loglik} : the residual log-likelihood for each model
\item \code{full.loglik} : the full log-likelihood for each model
\item \code{p} :  the number of fixed effects parameters for each model
\item \code{q} : the number of (non-zero) variance parameters for each model.
\item \code{b} : the number of variance parameters that are fixed or on the
boundary.  These parameters are not counted in the AIC or BIC.
\item \code{AIC} : the AIC for each model
\item \code{BIC} : the BIC for each model
\item \code{logdet} : the log-determinant used in adjusting the residual
log-likelihood for each model
}
}
\description{
icREML
}
\details{
This function calculates the AIC and BIC for each fitted asreml model in a
list.  Options are set and each model is updated.  The elements required
for the AIC and BIC are then calculated.
}
