% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{igrep}
\alias{igrep}
\title{Internal splinetools function.}
\usage{
igrep(patterns, x, ...)
}
\arguments{
\item{patterns}{list of character vectors.}

\item{x}{object for which patterns are sought.}

\item{\ldots}{Additional parameters.}
}
\value{
The index of matches of patterns in x.
}
\description{
igrep(patterns, x, ...)
lgrep(patterns, x, ...)
}
\author{
Julian Taylor \href{mailto:julian.taylor@adelaide.edu.au}{julian.taylor@adelaide.edu.au}
}
