% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{igrep}
\alias{igrep}
\alias{lgrep}
\title{Search for patterns in lists}
\usage{
igrep(patterns, x, ...)

lgrep(patterns, x, ...)
}
\arguments{
\item{patterns}{list of character vectors.}

\item{x}{object for which patterns are sought.}

\item{...}{Additional parameters.}
}
\value{
\code{lgrep}: The index of matches of the intersection of patterns in \code{x}.
\code{igrep}: The index of matches of the patterns in \code{x}.
}
\description{
Search for patterns in lists
}
