% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2sparse.R
\name{mat2sparse}
\alias{mat2sparse}
\title{mat2sparse}
\usage{
mat2sparse(mat)
}
\arguments{
\item{mat}{A symmetric matrix.}
}
\value{
A matrix with row, column and values from the lower triangle
of the matrix \code{mat}, which is row-major order required by
\code{asreml}.
}
\description{
Convert a symmetric matrix to a three column (matrix) sparse form
}
