% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{merge_formula}
\alias{merge_formula}
\title{Merge two formulae}
\usage{
merge_formula(form1, form2, LHS = TRUE, ...)
}
\arguments{
\item{form1}{a formula}

\item{form2}{a formula}

\item{LHS}{a logical scalar that indicates if there is a response in \code{form1}
and \code{form2}}

\item{\code{...}}{additional arguments}
}
\value{
A single model formula
}
\description{
Merge two formulae into a single formula
}
\details{
The function takes two formulae and combines them into a single formula
}
