% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvar1init.R
\name{mvar1init}
\alias{mvar1init}
\title{mvar1 initial values}
\usage{
mvar1init(
  gam,
  init.var,
  spatial = list("Run", "Range"),
  Traits = NULL,
  n.trait = length(init.var)
)
}
\arguments{
\item{gam}{a vparameters.table object from an \code{asreml}
call using \code{start.values=TRUE}}

\item{init.var}{initial variances for each level of the factor
given by the \code{Trait} argument with a names attribute that
matches \code{Traits}(see below)}

\item{spatial}{a list of character strings of the model for the spatial
effects}

\item{Traits}{a vaector of character strings giving the name of the traits
that defines the mvar1 structure in addition to rows and columns}

\item{n.trait}{the number of levels of the trait factor that
should match the length of the initial variances specified
in the argument \code{init.var}}
}
\value{
a vparamters.table object with initial values of the
mvar1 parameters set
}
\description{
mvar1 initial value function: mvar1init
}
\details{
This function sets up starting values for the parameters
of a multivariate autoregressive process of order 1
}
\author{
Ari Verbyla (averbyla at avdataanalytics.com.au)
}
