% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{Ops.formula}
\alias{Ops.formula}
\title{Addition operator for formulae}
\usage{
\method{Ops}{formula}(e1, e2)
}
\arguments{
\item{e1}{and e2 Two model formulae}
}
\value{
The sum of the two model formulae
}
\description{
An addition operator that works for formulae
}
