% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepData.R
\docType{data}
\name{prepData}
\alias{prepData}
\title{prepData: wheat multi-environment trials}
\description{
Masked data set.  Four sites of wheat variety trials all designed as
partially replicated designs, with replicaton being close to 50\\%.
The variables are as follows:
}
\details{
\itemize{
\item Site.  the site factor.
\item Line.  The variety in a plot.
\item Rep.  The rep a plot is in.
\item Row.  The row a plot is in.
\item Column.  The column that a plot is in.
\item yield.  The yield in tonnes per hectare.
\item Block.  The block a plot is in.
\item PhenoGroup.  A factor that indicates which phenology group a
line is in (only avaliable for sites 3 and 4 which were in year 2 of
the trials).
}
}
