% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage1.R
\name{stage1}
\alias{stage1}
\title{Stage 1 analysis function using asreml}
\usage{
stage1(
  fixed,
  random,
  data,
  residual = NULL,
  sparse = NULL,
  family = NULL,
  specific = list(),
  Genetic = "id",
  predGenetic = list(id = Genetic, classify = Genetic, rm = NULL),
  predict.options = NULL,
  type = "random",
  Trait = "Trait",
  n.trait = 1,
  keep.models = FALSE,
  mvar1.init.obj = NULL,
  s1trace = FALSE,
  ...
)
}
\arguments{
\item{fixed}{A model formula for the fixed effects passed to
asreml}

\item{random}{A model formula for the random effects passed to
asreml}

\item{data}{The data frame or data table containing the data to
be used in analysis.}

\item{residual}{A model formula for the residual effects passed to
asreml}

\item{sparse}{A model formula for the sparse effects passed to
asreml}

\item{family}{A family call passed to asreml}

\item{specific}{A list of terms to be added to the model that are
specific to Trait values.  Each component of the list is
itself a list with the type of effect to be included as a
component (fixed or random), and a vector specifying which
Trait values require the term as another component.  This
component can have names (character vector) or be numeric.  In
the latter case the values are assumed to correspond to the
order of the Trait factor levels.}

\item{Genetic}{A character vector specifying the name of the
genetic effect that is being fitted, The default is "id".}

\item{predGenetic}{A list specifying the \code{id} argument which
gives the name of the genetic term to be fitted at stage 2
(default, the same as the \code{Genetic} argument, the
\code{classify} term for prediction of genetic effects
(default, the same as the \code{Genetic} argument), and any
terms to remove, called \code{rm} in the list, from the model
for prediction of fixed genetic effects.  The default for the
classify is set to the value of the \code{Genetic} argument by
default, while \code{rm} is initially set to \code{NULL}}

\item{predict.options}{A list of options for the prediction step
in the stage 1 analyses, that are applicable to all stage 1
analyses.}

\item{type}{A character vector specifying the type of interaction
between Trait and Genetic.  The default is "random".  The
alternative is "fixed".}

\item{Trait}{A character string that defines the trait involved.}

\item{n.trait}{The number of levels of Trait to be used in the
first stage analysis.  The default is 1.}

\item{keep.models}{A logical that flags whether all the models
fitted at stage 1 should be saved as part of the output list.
If \code{type} is random, the returned models have random
Genetic effects, otherwise fixed Genetic effects.}

\item{mvar1.init.obj}{A list of size 3.  The component \code{val}
is a vector of initial values (variances) for
each of the levels of Trait.  The other two components are
\code{Run} and \code{Range} which are characters that specify
the factors for the field Run and Range.  If NULL, the
multivariate ar1 is not being fitted and this is the default.
This structure will only be used for repeated measures data.}

\item{s1trace}{A logical scalar indicating if trace printing is to
be switched on for debugging}

\item{...}{Further arguments to be used in the underlying
\code{asreml} fit, for example \code{workspace}, \code{maxit}
that are standard asreml arguments.}
}
\value{
A list of named components.  \code{call} contains the call
to this function. \code{Vt} contains a three column data frame
of the variance matrix of the first stage effects, row,
column, and value.  It has two attributes, namely
\code{"INVERSE"} which is \code{FALSE}, and \code{"rowNames"}
required for \code{vm} in the second stage analysis in
\code{asreml}.  \code{bdVt} is a sparse block diagonal
variance matrix for use in the second stage analysis using
\code{vm} in package \code{asreml}. \code{V} and \code{W} are
lists of the variance and weight matrices from the stage 1
analyses. \code{pred.df} is a data frame with Trait by genetic
effects for use in the second stage analysis.  \code{type}
records the type of first stage analysis used (fixed or
random).  \code{vc} is a list of estimated variance parameters
for each stage 1 analysis.  \code{vcMat} is a list of matrices
of variance parameters for all stage 1 models. \code{models}
is a list of the fitted model objects from the stage 1 fits if
\code{keep.models} has been set to \code{TRUE}.
}
\description{
A function for stage 1 analyses of a two stage analysis
}
\details{
A function that calls asreml for each component of a stage 1
analysis.
}
