% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genherit1.R
\name{genherit1}
\alias{genherit1}
\title{generic genherit1}
\usage{
genherit1(asr, ...)
}
\arguments{
\item{asr}{A model object.}

\item{...}{Other arguments to \code{asreml} can be given.}

\item{id}{A character string giving the name of the genetic effect.}

\item{only}{A character vector specifying which variables form the prediction model. The default value is \code{id}.}

\item{Gmat}{A genetic variance or inverse variance matrix.  It must have a
logical attribute called \code{"INVERSE"}.  If the attribute is \code{TRUE},
\code{Gmat} is an inverse variance matrix.  If the attribute is \code{FALSE},
\code{Gmat} is a variance matrix.  The default for \code{Gmat} is \code{NULL}
in which case the variance matrix is assumed to be the scaled identity.}
}
\value{
The estimation of generalised heritability (a number between 0 and 1)
}
\description{
Generic function for Generalised Heritability for a single trial
}
\details{
Estimates the generalized heritability of one trial
}
\references{
Oakey, H., Verbyla, A. P., Cullis, B. R., Wei, X., & Pitchford, W. S. (2007). Joint
modeling of additive and non-additive genetic line effects in multi-environment trials.
Theoretical and Applied Genetics, 114(8), 1319-1332.
}
\author{
Ari Verbyla \email{averbyla at avdataanalytics.com.au}
}
