% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{lgrep}
\alias{lgrep}
\title{Internal lmmtools function.}
\usage{
lgrep(patterns, x, ...)
}
\arguments{
\item{...}{Additional parameters.}

\item{patterns:}{list of character vectors.}

\item{x:}{object for which patterns are sought.}
}
\value{
The index of matches of the intersection of patterns in x.
}
\description{
lmmtools internal functions
}
\details{
These functions are used internally in lmmtools.

igrep(patterns, x, ...)
lgrep(patterns, x, ...)
}
\author{
Ari Verbyla (averbyla at avdataanalytics.com.au)
}
