% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2df.R
\name{mat2df}
\alias{mat2df}
\title{mat2df}
\usage{
mat2df(mat)
}
\arguments{
\item{mat}{A symmetric matrix.}
}
\value{
A data frame with row, column and values from the lower triangle
of the matrix \code{mat}
}
\description{
Convert a symmetric matrix to a three column (matrix) sparse form
}
